windbird/requirements.txt
# 基础框架
Django==5.1.6
djangorestframework==3.15.2

# 数据库
mysqlclient==2.2.1

# 安全
certifi==2025.1.31
urllib3==2.3.0

# 工具
python-dotenv==1.0.0
gunicorn==21.2.0

# 网络请求
requests==2.32.3
charset-normalizer==3.4.1
idna==3.10

# 缓存
django-redis==5.3.0

# 其他
asgiref==3.8.1
sqlparse==0.5.3 